WEBAVS-1876 Les traitements prévisionnnels du batch contentieux pour les sommations de type "standards" et "bulletins neutres" se terminent en échec

+++ webavs-web/src/main/java/globaz/osiris/db/utils/CAReferenceBVR.java
@@ -199,10 +199,8 @@ public class CAReferenceBVR {
         int valIdRole = Integer.parseInt(CAApplication.getApplicationOsiris().getProperty(
                 IntReferenceBVRParser.VAL_ID_ROLE));
 
-        // Modification pour le BMS uniquement !
         String sIdExterneRole = removeNotLetterNotDigit(idExterneRole.split("-")[0]);
-        int iIdExterneRole = Integer.parseInt(sIdExterneRole);
-        // sIdExterneRole = StringUtils.leftPad(String.valueOf(iIdExterneRole), getLengthRefDebiteur(), '0');
+        long iIdExterneRole = Long.parseLong(sIdExterneRole);
 
         String role = JadeStringUtil.rightJustifyInteger(idRole, getLengthIdRole());
         String refDebiteur = JadeStringUtil.rightJustifyInteger(String.valueOf(iIdExterneRole), getLengthRefDebiteur());
